pyproject.toml
[tool.poetry]
name = "abz-agency"
version = "0.1.0"
description = ""
authors = ["Sergey Radyukhin <22dollars@mail.ru>"]
readme = "README.md"
packages = [{include = "abz_agency"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.5"
python-dotenv = "^0.21.1"
django-extensions = "^3.2.1"
django-bootstrap5 = "^22.2"
django-bootstrap = "^0.2.4"
django-bootstrap4 = "^22.3"
psycopg2-binary = "^2.9.5"
django-mptt = "^0.14.0"


[tool.poetry.group.dev.dependencies]
ipython = "^8.9.0"
flake8 = "^6.0.0"
django-seed = "^0.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
